backend/requirements.txt
fastapi==0.0.16
uvicorn[standard]==0.35.0
httpx
python-dotenv